+++ FurnitureStoreLibrary/Furniture.cs
@@ -10,12 +10,12 @@ namespace FurnitureStore.Library
         public int furnitureID { get; set; }
         public string furnitureName { get; set; }
         public string furnitureType { get; set; }
-        public int furniturePrice { get; set; }
+        public float furniturePrice { get; set; }
         public string furniturePieces { get; set; }
         public string furnitureDescription { get; set; }
 
 
-        public Furniture(int furnitureID, string furnitureName, string furnitureType, int furniturePrice, string furniturePieces, string furnitureDescription)
+        public Furniture(int furnitureID, string furnitureName, string furnitureType, float furniturePrice, string furniturePieces, string furnitureDescription)
         {
             this.furnitureID = furnitureID;
             this.furnitureName = furnitureName;
